第一次提交

+++ src/main/java/com/sweet/modular/stock/StockDto.java
@@ -12,5 +12,5 @@ import java.math.BigDecimal;
 public class StockDto {
     private Long id;
     private String name;
-    private BigDecimal count;
+    private Long count;
 }
